refactor(notification): address PR review feedback (#1693)
