+++ src/core/pipes/notification/notification.normalizer.ts
@@ -5,21 +5,7 @@ import { getBusinessKey } from '@/core/models/notification/notification.helpers'
 import type { NotificationPreferences } from '@/core/stores/settings/settings.types';
 import { NotificationType } from '@/core/models/notification/notification.types';
 import type { FlatNotification } from '@/core/models/notification/notification.types';
-
-// Maps each NotificationType enum to its corresponding NotificationPreferences key.
-// Needed because NotificationType values don't match the preference key names directly.
-// exported for tests
-export const NOTIFICATION_TYPE_TO_PREFERENCE_KEY: Record<NotificationType, keyof NotificationPreferences> = {
-  [NotificationType.Follow]: 'follow',
-  [NotificationType.NewFriend]: 'newFriend',
-  [NotificationType.TagPost]: 'tagPost',
-  [NotificationType.TagProfile]: 'tagProfile',
-  [NotificationType.Reply]: 'reply',
-  [NotificationType.Repost]: 'repost',
-  [NotificationType.Mention]: 'mention',
-  [NotificationType.PostDeleted]: 'postDeleted',
-  [NotificationType.PostEdited]: 'postEdited',
-};
+import { NOTIFICATION_TYPE_TO_PREFERENCE_KEY } from './notification.constants';
 
 export class NotificationNormalizer {
   private constructor() {}
